+++ evm/src/interfaces/IHubPortal.sol
@@ -23,14 +23,14 @@ interface IHubPortal is IPortal {
     /// @param  destinationChainId The chain Id of the destination chain.
     /// @param  index              The the M token index.
     /// @param  messageId          The unique ID of the sent message.
-    event MTokenIndexSent(uint256 destinationChainId, uint128 index, bytes32 messageId);
+    event MTokenIndexSent(uint32 destinationChainId, uint128 index, bytes32 messageId);
 
     /// @notice Emitted when the Registrar key is sent to a destination chain.
     /// @param  destinationChainId The chain Id of the destination chain.
     /// @param  key                The key that was sent.
     /// @param  value              The value that was sent.
     /// @param  messageId          The unique ID of the sent message.
-    event RegistrarKeySent(uint256 destinationChainId, bytes32 indexed key, bytes32 value, bytes32 messageId);
+    event RegistrarKeySent(uint32 destinationChainId, bytes32 indexed key, bytes32 value, bytes32 messageId);
 
     /// @notice Emitted when the Registrar list status for an account is sent to a destination chain.
     /// @param  destinationChainId The chain Id of the destination chain.
@@ -39,7 +39,7 @@ interface IHubPortal is IPortal {
     /// @param  status             The status of the account in the list.
     /// @param  messageId          The unique ID of the sent message.
     event RegistrarListStatusSent(
-        uint256 destinationChainId, bytes32 indexed listName, address indexed account, bool status, bytes32 messageId
+        uint32 destinationChainId, bytes32 indexed listName, address indexed account, bool status, bytes32 messageId
     );
 
     ///////////////////////////////////////////////////////////////////////////
@@ -76,7 +76,7 @@ interface IHubPortal is IPortal {
     /// @param  destinationChainId The chain Id of the destination chain.
     /// @param  refundAddress      The refund address to receive excess native gas.
     /// @return messageId          The ID uniquely identifying the message.
-    function sendMTokenIndex(uint256 destinationChainId, bytes32 refundAddress) external payable returns (bytes32 messageId);
+    function sendMTokenIndex(uint32 destinationChainId, bytes32 refundAddress) external payable returns (bytes32 messageId);
 
     /// @notice Sends the Registrar key to the destination chain.
     /// @param  destinationChainId The chain Id of the destination chain.
@@ -84,7 +84,7 @@ interface IHubPortal is IPortal {
     /// @param  refundAddress      The refund address to receive excess native gas.
     /// @return messageId          The ID uniquely identifying the message.
     function sendRegistrarKey(
-        uint256 destinationChainId,
+        uint32 destinationChainId,
         bytes32 key,
         bytes32 refundAddress
     ) external payable returns (bytes32 messageId);
@@ -96,7 +96,7 @@ interface IHubPortal is IPortal {
     /// @param  refundAddress      The refund address to receive excess native gas.
     /// @return messageId          The ID uniquely identifying the message.
     function sendRegistrarListStatus(
-        uint256 destinationChainId,
+        uint32 destinationChainId,
         bytes32 listName,
         address account,
         bytes32 refundAddress
